+++ src/components/Teaser/Body.jsx
@@ -7,10 +7,18 @@ const TeaserBody = (props) => {
   const { variation, data } = props;
 
   const hasType = data.href?.[0]?.['@type'];
-  const variationComponent = hasType && `Teaser|${data.href?.[0]?.['@type']}`;
+  const legacyVariationComponentName =
+    hasType && `Teaser|${data.href?.[0]?.['@type']}`;
 
+  // Compatible with the previous version of the component registry
+  // and the Volto 16 one.
   const BodyComponent =
-    config.resolve(variationComponent).component ||
+    (config?.resolve &&
+      config?.resolve(legacyVariationComponentName)?.component) ||
+    (config?.getComponent &&
+      hasType &&
+      config.getComponent({ name: 'Teaser', dependencies: [hasType] })
+        .component) ||
     variation?.template ||
     DefaultBody;